+++ src/functions/public/ConvertFrom-ResultPropertyValueCollectionToString.ps1
@@ -20,6 +20,7 @@ function ConvertFrom-ResultPropertyValueCollectionToString {
     #>
 
     param (
+        # ResultPropertyValueCollection object to convert to a string
         [System.DirectoryServices.ResultPropertyValueCollection]$ResultPropertyValueCollection
     )