+++ src/functions/public/Get-DirectoryEntry.ps1
@@ -44,6 +44,7 @@ function Get-DirectoryEntry {
         # Properties of the target object to retrieve
         [string[]]$PropertiesToLoad,
 
+        # Mapping of SID types to descriptions used for converting security identifiers
         [hashtable]$SidTypeMap = (Get-SidTypeMap),
 
         # In-process cache to reduce calls to other processes or to disk
@@ -105,7 +106,8 @@ function Get-DirectoryEntry {
                     $($Credential.GetNetworkCredential().password)
                 )
 
-            } else {
+            }
+            else {
                 $DirectoryEntry = [System.DirectoryServices.DirectoryEntry]::new($DirectoryPath)
             }
 
@@ -118,7 +120,8 @@ function Get-DirectoryEntry {
             $DirectoryEntry |
             Add-Member -MemberType NoteProperty -Name 'Domain' -Value $SampleUser.Domain -Force
 
-        } else {
+        }
+        else {
 
             # Otherwise the DirectoryPath is an LDAP path or a WinNT path (treated the same at this stage)
             Write-LogMsg @Log -Text "[System.DirectoryServices.DirectoryEntry]::new('$DirectoryPath')"
@@ -129,7 +132,8 @@ function Get-DirectoryEntry {
                     $($Credential.UserName),
                     $($Credential.GetNetworkCredential().password)
                 )
-            } else {
+            }
+            else {
                 $DirectoryEntry = [System.DirectoryServices.DirectoryEntry]::new($DirectoryPath)
             }
 
@@ -144,7 +148,8 @@ function Get-DirectoryEntry {
             # If the $DirectoryPath was invalid, this line will return an error
             $null = $DirectoryEntry.RefreshCache($PropertiesToLoad)
 
-        } catch {
+        }
+        catch {
 
             $Log['Type'] = 'Warning' # PS 5.1 can't override the Splat by calling the param, so we must update the splat manually