+++ src/functions/public/Get-KnownSidByName.ps1
@@ -1,6 +1,7 @@
 function Get-KnownSidByName {
 
     param (
+        # Hashtable containing well-known SIDs as keys with their properties as values
         [hashtable]$WellKnownSIDBySID
     )