+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [4.0.545] - 2025-05-17 - add param descriptions
+
+## [4.0.544] - 2025-05-17 - add param descriptions
+
 ## [4.0.543] - 2025-05-17 - update psakefile
 
 ## [4.0.542] - 2025-05-17 - update psakefile